clean up conduits, move rejuvenating wind to shared talent
