+++ src/analysis/retail/hunter/shared/index.ts
@@ -14,15 +14,7 @@ export { default as NaturalMending } from './talents/NaturalMending';
 export { default as SpellFocusCost } from './SpellFocusCost';
 export { default as Trailblazer } from './talents/Trailblazer';
 export { default as TranquilizingShot } from './spells/TranquilizingShot';
-export { default as EnfeebledMark } from './conduits/kyrian/EnfeebledMark';
-export { default as NecroticBarrage } from './conduits/necrolord/NecroticBarrage';
-export { default as SpiritAttunement } from './conduits/nightfae/SpiritAttunement';
-export { default as EmpoweredRelease } from './conduits/venthyr/EmpoweredRelease';
-export { default as HarmonyOfTheTortollan } from './conduits/HarmonyOfTheTortollan';
-export { default as MarkmansAdvantage } from './conduits/MarkmansAdvantage';
-export { default as RejuvenatingWind } from './conduits/RejuvenatingWind';
-export { default as ResilienceOfTheHunter } from './conduits/ResilienceOfTheHunter';
-export { default as ReversalOfFortune } from './conduits/ReversalOfFortune';
+export { default as RejuvenatingWind } from './talents/RejuvenatingWind';
 export { default as DeathChakrams } from './talents/DeathChakrams';
 export { default as WailingArrow } from './talents/WailingArrow';
 export { default as WailingArrowPrepullNormalizer } from './talents/WailingArrowPrepullNormalizer';
